requirements.txt
# CyberAI Sentinel - Dependencias del Proyecto
# Desarrollado por: Vicente Alonso
# Especialista en Ciberseguridad y Administración de Sistemas

# =============================================================================
# CORE FRAMEWORKS
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1

# =============================================================================
# MACHINE LEARNING & AI
# =============================================================================
tensorflow>=2.20.0
torch>=2.1.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# =============================================================================
# DATA PROCESSING & ANALYSIS
# =============================================================================
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
statsmodels==0.14.0
networkx==3.2.1

# =============================================================================
# DATABASE & CACHE
# =============================================================================
psycopg2-binary==2.9.9
redis==5.0.1
elasticsearch==8.11.0
pymongo==4.6.0

# =============================================================================
# SECURITY & AUTHENTICATION
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8
hashlib-compat==1.0.1

# =============================================================================
# HTTP & API CLIENT
# =============================================================================
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# =============================================================================
# LOGGING & MONITORING
# =============================================================================
loguru==0.7.2
prometheus-client==0.19.0
structlog==23.2.0

# =============================================================================
# TESTING
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# =============================================================================
# CONFIGURATION & ENVIRONMENT
# =============================================================================
python-dotenv==1.0.0
pydantic-settings==2.1.0
click==8.1.7
typer==0.9.0

# =============================================================================
# SCHEDULING & TASKS
# =============================================================================
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# =============================================================================
# FILE PROCESSING
# =============================================================================
openpyxl==3.1.2
python-magic==0.4.27
pillow==10.1.0

# =============================================================================
# NETWORK & SECURITY
# =============================================================================
scapy==2.5.0
netaddr==0.10.1
ipaddress==1.0.23

# =============================================================================
# OPTIONAL: ADVANCED FEATURES
# =============================================================================
# jupyter==1.0.0
# streamlit==1.28.1
# gradio==4.8.0
# dash==2.14.2
